Fix panel shade position
Panels like the [fuzzy finder](https://atom.io/packages/fuzzy-finder) often cover the rest of the UI in a "shade" to help highlight the panel. In this theme, the shade was implemented as an ::after element with position absolute and negative insets to extend it beyond the screen.

However, its parent was position relative. An element with position absolute inside an element with position relative is relative to its parent. So the insets were sticking out only a little beyond the panel itself.

This change removes the whole atom-panel.modal:after selector because the default atom behavior seems to be what the theme was going for in the first place.

Fixes #1